refactor: Code splitting for input types, output types and exp filter
